Quellcode/requirements.txt
tensorflow-gpu
argparse
dill
fire
matplotlib
psutil
requests
scikit-image
scipy
slidingwindow
tqdm
git+https://github.com/ppwwyyxx/tensorpack.git
tensorflow-hub
numpy
sklearn